# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Travel Mate API",
    "version": "1.0.0",
    "description": "Dokumentacja API dla aplikacji Travel Mate - kompleksowej platformy do planowania podróży",
    "contact": {
      "name": "Travel Mate Team",
      "email": "contact@travelmate.com"
    },
    "license": {
      "name": "Private",
      "url": "https://travelmate.com/license"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token z Firebase Authentication. Otrzymasz go po zalogowaniu przez POST /api/auth/login"
      }
    },
    "schemas": {
...
      "Flight": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "airline": {
            "type": "string"
          },
          "stops": {
            "type": "string"
          },
          "departureTime": {
            "type": "string"
          },
          "departureAirport": {
            "type": "string"
          },
          "arrivalTime": {
            "type": "string"
          },
          "arrivalAirport": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "returnAirline": {
            "type": "string"
          },
          "returnStops": {
            "type": "string"
          },
          "returnDepartureTime": {
            "type": "string"
          },
          "returnDepartureAirport": {
            "type": "string"
          },
          "returnArrivalTime": {
            "type": "string"
          },
          "returnArrivalAirport": {
            "type": "string"
          },
          "returnDuration": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  },
  "security": [],
...
  "paths": {
...
    "/api/flights/search": {
      "get": {
        "tags": [
          "Flights"
        ],
        "summary": "Wyszukaj loty",
        "description": "Wyszukuje dostępne loty za pomocą Amadeus API. Obsługuje loty w jedną stronę i powrotne.",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Kod IATA lotniska wylotu (np. WAW)",
            "example": "WAW"
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Kod IATA lotniska przylotu (np. LHR)",
            "example": "LHR"
          },
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Data wylotu (YYYY-MM-DD)",
            "example": "2026-06-15"
          },
          {
            "in": "query",
            "name": "returnDate",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Data powrotu (YYYY-MM-DD) - opcjonalna dla lotów powrotnych",
            "example": "2026-06-22"
          },
          {
            "in": "query",
            "name": "passengers",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Liczba pasażerów",
            "example": 2
          }
        ],
        "responses": {
          "200": {
            "description": "Lista dostępnych lotów",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "flights": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Flight"
                      }
                    }
                  }
                },
                "example": {
                  "success": true,
                  "flights": [
                    {
                      "id": "1",
                      "airline": "LOT Polish Airlines",
                      "stops": "0",
                      "departureTime": "10:30",
                      "departureAirport": "WAW",
                      "arrivalTime": "12:45",
                      "arrivalAirport": "LHR",
                      "duration": "2h 15m",
                      "type": "round-trip",
                      "price": 450,
                      "returnAirline": "LOT Polish Airlines",
                      "returnStops": "0",
                      "returnDepartureTime": "14:00",
                      "returnDepartureAirport": "LHR",
                      "returnArrivalTime": "18:15",
                      "returnArrivalAirport": "WAW",
                      "returnDuration": "2h 15m"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Nieprawidłowe parametry wyszukiwania",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Parametry from, to i date są wymagane"
                }
              }
            }
          },
          "500": {
            "description": "Błąd serwera lub Amadeus API",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}